worker/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
SQLAlchemy==2.0.22
redis==5.0.8
pika==1.3.1
prometheus-client==0.16.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0

